backend/pyproject.toml
[tool.poetry]
name = "tars-server"
version = "2.0.0"
description = "Twitcasting Autonomous Recording System"
authors = ["clustlight <git@clustlight.net>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
websockets = "^12.0"
sqlalchemy = "^2.0.25"
uvicorn = "^0.27.1"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
